Move ReaddirHandle into its own file
`inode.rs` is getting too long to be readable.
